training/requirements.txt
torch>=2.0.0
numpy>=1.24.0
python-chess>=1.999
tqdm>=4.65.0
tensorboard>=2.14.0
onnx>=1.14.0
onnxruntime>=1.16.0
